refactor(track-ct): address PR #16 review comments — split provisioner, extract service, rename
Implement all 6 inline review comments from jrosseel on PR #16:

1. provisioner.ts split: SharedClusterProvisioner → shared-cluster.provisioner.ts,
   ExternalWebhookProvisioner → external-webhook.provisioner.ts,
   _ReadExternalWebhookConfig → external-webhook.config.ts.

2. provisioner.types.ts single-type: moved ExternalWebhookProvisionerConfig
   to external-webhook.provisioner.types.ts, leaving only ClusterTenantProvisioner.

3. registry.ts factory-only: DefaultClusterTenantProvisionerRegistry class +
   RegisteredProvisioner interface moved to registry.infra.ts; factory stays.

4. cluster-tenants.ts thin routes: all mapping/validation/projection helpers
   extracted to cluster-tenants.service.ts (_PascalCase exports).

5. cluster-tenants.types.ts → cluster-tenants.models.ts: request/response
   interfaces renamed to models per reviewer feedback.

6. _K8sApplyResource → __K8sApplyResource: renamed in infra/k8s.ts and all
   four operator consumers (package-wide scope warrants __ prefix).

Test imports updated to reflect new file locations.
