+++ apps/control-plane/src/routes/cluster-tenants.ts
@@ -1,132 +1,10 @@
 import { Router } from "express";
-import { ClusterTenantComputeMode, ClusterTenantIsolationTier, ClusterTenantPhase, ClusterTenantTierUnavailableCode } from "@opencrane/contracts";
-import type { ClusterTenant, ClusterTenantProvisionerRegistry, ClusterTenantResourceQuota } from "@opencrane/contracts";
+import { ClusterTenantPhase, ClusterTenantTierUnavailableCode } from "@opencrane/contracts";
+import type { ClusterTenantProvisionerRegistry } from "@opencrane/contracts";
 import type { Prisma, PrismaClient } from "@prisma/client";
 
-import type { ClusterTenantComputeInput, ClusterTenantCreateRequest, ClusterTenantResourcesInput, ClusterTenantUpdateRequest } from "./cluster-tenants.types.js";
-
-/** A cluster_tenants row as read back from Prisma (subset consumed here). */
-type ClusterTenantRow = Prisma.ClusterTenantGetPayload<Record<string, never>>;
-
-/** Map the contract isolation tier (lowercase) to the Prisma enum member (PascalCase). */
-function _toPrismaTier(tier: ClusterTenantIsolationTier): "Shared" | "DedicatedNodes" | "DedicatedCluster"
-{
-  switch (tier)
-  {
-    case ClusterTenantIsolationTier.Shared: return "Shared";
-    case ClusterTenantIsolationTier.DedicatedNodes: return "DedicatedNodes";
-    case ClusterTenantIsolationTier.DedicatedCluster: return "DedicatedCluster";
-  }
-}
-
-/** Map the contract compute mode (lowercase) to the Prisma enum member (PascalCase). */
-function _toPrismaCompute(mode: ClusterTenantComputeMode): "Shared" | "Dedicated"
-{
-  return mode === ClusterTenantComputeMode.Dedicated ? "Dedicated" : "Shared";
-}
-
-/** Map a Prisma isolation-tier enum member back to the contract tier value. */
-function _fromPrismaTier(value: string): ClusterTenantIsolationTier
-{
-  switch (value)
-  {
-    case "DedicatedNodes": return ClusterTenantIsolationTier.DedicatedNodes;
-    case "DedicatedCluster": return ClusterTenantIsolationTier.DedicatedCluster;
-    default: return ClusterTenantIsolationTier.Shared;
-  }
-}
-
-/** Map a Prisma compute-mode enum member back to the contract compute value. */
-function _fromPrismaCompute(value: string): ClusterTenantComputeMode
-{
-  return value === "Dedicated" ? ClusterTenantComputeMode.Dedicated : ClusterTenantComputeMode.Shared;
-}
-
-/** Map the stored phase string back to the contract phase enum (defaults to pending). */
-function _fromPrismaPhase(value: string): ClusterTenantPhase
-{
-  switch (value)
-  {
-    case "provisioning": return ClusterTenantPhase.Provisioning;
-    case "ready": return ClusterTenantPhase.Ready;
-    case "failed": return ClusterTenantPhase.Failed;
-    default: return ClusterTenantPhase.Pending;
-  }
-}
-
-/** Whether a value is one of the contract isolation-tier strings. */
-function _isIsolationTier(value: unknown): value is ClusterTenantIsolationTier
-{
-  return value === ClusterTenantIsolationTier.Shared || value === ClusterTenantIsolationTier.DedicatedNodes || value === ClusterTenantIsolationTier.DedicatedCluster;
-}
-
-/** Whether a value is one of the contract compute-mode strings. */
-function _isComputeMode(value: unknown): value is ClusterTenantComputeMode
-{
-  return value === ClusterTenantComputeMode.Shared || value === ClusterTenantComputeMode.Dedicated;
-}
-
-/**
- * Validate a compute block: a dedicated mode requires a node pool, otherwise the
- * operator could place pods on no machines at all.
- *
- * @param compute - Compute input from the request body.
- * @returns A validation error message, or null when valid.
- */
-function _validateCompute(compute: ClusterTenantComputeInput | undefined): string | null
-{
-  if (!compute || !_isComputeMode(compute.mode))
-  {
-    return "compute.mode must be 'shared' or 'dedicated'.";
-  }
-  if (compute.mode === ClusterTenantComputeMode.Dedicated && !compute.nodePool?.trim())
-  {
-    return "compute.nodePool is required when compute.mode is 'dedicated'.";
-  }
-  return null;
-}
-
-/**
- * Validate a resources block: the quota object must be present (it is the
- * resource ceiling enforced over the customer's namespace).
- *
- * @param resources - Resources input from the request body.
- * @returns A validation error message, or null when valid.
- */
-function _validateResources(resources: ClusterTenantResourcesInput | undefined): string | null
-{
-  if (!resources || typeof resources.quota !== "object" || resources.quota === null)
-  {
-    return "resources.quota must be provided.";
-  }
-  return null;
-}
-
-/**
- * Project a stored row into the shared {@link ClusterTenant} contract shape.
- *
- * @param row - The persisted cluster_tenants row.
- * @returns The contract representation returned to API clients.
- */
-function _toContract(row: ClusterTenantRow): ClusterTenant
-{
-  return {
-    name: row.name,
-    displayName: row.displayName,
-    isolationTier: _fromPrismaTier(row.isolationTier as unknown as string),
-    compute: {
-      mode: _fromPrismaCompute(row.computeMode as unknown as string),
-      ...(row.nodePool ? { nodePool: row.nodePool } : {}),
-    },
-    resources: { quota: (row.quota as ClusterTenantResourceQuota | null) ?? {} },
-    status: {
-      phase: _fromPrismaPhase(row.phase),
-      ...(row.message ? { message: row.message } : {}),
-      ...(row.boundNamespace ? { boundNamespace: row.boundNamespace } : {}),
-      ...(row.provisioner ? { provisioner: row.provisioner } : {}),
-    },
-  };
-}
+import type { ClusterTenantCreateRequest, ClusterTenantUpdateRequest } from "./cluster-tenants.models.js";
+import { _IsIsolationTier, _ToContract, _ToPrismaCompute, _ToPrismaTier, _ValidateCompute, _ValidateResources } from "./cluster-tenants.service.js";
 
 /**
  * CRUD router for the first-class cluster tenant (customer / isolation unit).
@@ -147,7 +25,7 @@ export function clusterTenantsRouter(prisma: PrismaClient, registry: ClusterTena
   router.get("/", async function _listClusterTenants(req, res)
   {
     const rows = await prisma.clusterTenant.findMany({ orderBy: { createdAt: "asc" } });
-    res.json(rows.map(_toContract));
+    res.json(rows.map(_ToContract));
   });
 
   /** Get a single cluster tenant by name. */
@@ -159,7 +37,7 @@ export function clusterTenantsRouter(prisma: PrismaClient, registry: ClusterTena
       res.status(404).json({ error: "Cluster tenant not found", code: "CLUSTER_TENANT_NOT_FOUND" });
       return;
     }
-    res.json(_toContract(row));
+    res.json(_ToContract(row));
   });
 
   /** Get just the observed status of a cluster tenant. */
@@ -171,7 +49,7 @@ export function clusterTenantsRouter(prisma: PrismaClient, registry: ClusterTena
       res.status(404).json({ error: "Cluster tenant not found", code: "CLUSTER_TENANT_NOT_FOUND" });
       return;
     }
-    res.json(_toContract(row).status);
+    res.json(_ToContract(row).status);
   });
 
   /** Create a cluster tenant, gating the requested isolation tier on the registry. */
@@ -186,7 +64,7 @@ export function clusterTenantsRouter(prisma: PrismaClient, registry: ClusterTena
       res.status(400).json({ error: "name and displayName are required.", code: "VALIDATION_ERROR" });
       return;
     }
-    if (!_isIsolationTier(body.isolationTier))
+    if (!_IsIsolationTier(body.isolationTier))
     {
       res.status(400).json({ error: "isolationTier must be 'shared', 'dedicatedNodes', or 'dedicatedCluster'.", code: "VALIDATION_ERROR" });
       return;
@@ -194,13 +72,13 @@ export function clusterTenantsRouter(prisma: PrismaClient, registry: ClusterTena
 
     // 2. Validate compute placement and resource gating — a dedicated pool needs
     //    a node-pool name, and a quota ceiling must always be supplied.
-    const computeError = _validateCompute(body.compute);
+    const computeError = _ValidateCompute(body.compute);
     if (computeError)
     {
       res.status(400).json({ error: computeError, code: "VALIDATION_ERROR" });
       return;
     }
-    const resourcesError = _validateResources(body.resources);
+    const resourcesError = _ValidateResources(body.resources);
     if (resourcesError)
     {
       res.status(400).json({ error: resourcesError, code: "VALIDATION_ERROR" });
@@ -220,14 +98,14 @@ export function clusterTenantsRouter(prisma: PrismaClient, registry: ClusterTena
       data: {
         name: body.name.trim(),
         displayName: body.displayName.trim(),
-        isolationTier: _toPrismaTier(body.isolationTier),
-        computeMode: _toPrismaCompute(body.compute.mode),
+        isolationTier: _ToPrismaTier(body.isolationTier),
+        computeMode: _ToPrismaCompute(body.compute.mode),
         nodePool: body.compute.nodePool?.trim() || null,
         quota: (body.resources.quota as Prisma.InputJsonValue),
         phase: ClusterTenantPhase.Pending,
       },
     });
-    res.status(201).json(_toContract(created));
+    res.status(201).json(_ToContract(created));
   });
 
   /** Update a cluster tenant, re-gating the isolation tier when it changes. */
@@ -258,7 +136,7 @@ export function clusterTenantsRouter(prisma: PrismaClient, registry: ClusterTena
     //    must not be moved to a tier no provisioner can serve.
     if (body.isolationTier !== undefined)
     {
-      if (!_isIsolationTier(body.isolationTier))
+      if (!_IsIsolationTier(body.isolationTier))
       {
         res.status(400).json({ error: "isolationTier must be 'shared', 'dedicatedNodes', or 'dedicatedCluster'.", code: "VALIDATION_ERROR" });
         return;
@@ -268,26 +146,26 @@ export function clusterTenantsRouter(prisma: PrismaClient, registry: ClusterTena
         res.status(422).json({ error: `No provisioner is registered for isolation tier '${body.isolationTier}'.`, code: ClusterTenantTierUnavailableCode });
         return;
       }
-      data.isolationTier = _toPrismaTier(body.isolationTier);
+      data.isolationTier = _ToPrismaTier(body.isolationTier);
     }
 
     // 3. Re-validate compute placement when changed (dedicated needs a pool).
     if (body.compute !== undefined)
     {
-      const computeError = _validateCompute(body.compute);
+      const computeError = _ValidateCompute(body.compute);
       if (computeError)
       {
         res.status(400).json({ error: computeError, code: "VALIDATION_ERROR" });
         return;
       }
-      data.computeMode = _toPrismaCompute(body.compute.mode);
+      data.computeMode = _ToPrismaCompute(body.compute.mode);
       data.nodePool = body.compute.nodePool?.trim() || null;
     }
 
     // 4. Re-validate the resources block when changed (quota must be present).
     if (body.resources !== undefined)
     {
-      const resourcesError = _validateResources(body.resources);
+      const resourcesError = _ValidateResources(body.resources);
       if (resourcesError)
       {
         res.status(400).json({ error: resourcesError, code: "VALIDATION_ERROR" });
@@ -297,7 +175,7 @@ export function clusterTenantsRouter(prisma: PrismaClient, registry: ClusterTena
     }
 
     const updated = await prisma.clusterTenant.update({ where: { name: req.params.name }, data });
-    res.json(_toContract(updated));
+    res.json(_ToContract(updated));
   });
 
   /** Delete a cluster tenant. */
